third_party/WebKit/LayoutTests/accessibility/textarea-caret-position-expected.txt

  
This tests that caret position is reported correctly for textarea elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textareaAccessible.selectionStart is 0
PASS textareaAccessible.selectionEnd is 0
PASS textareaAccessible.selectionStart is 7
PASS textareaAccessible.selectionEnd is 7
PASS textareaAccessible.selectionStart is 14
PASS textareaAccessible.selectionEnd is 14
PASS textareaAccessible.selectionStart is 14
PASS textareaAccessible.selectionEnd is 14
PASS emptyTextareaAccessible.selectionStart is 0
PASS emptyTextareaAccessible.selectionEnd is 0
PASS emptyTextareaAccessible.selectionStart is 0
PASS emptyTextareaAccessible.selectionEnd is 0
PASS successfullyParsed is true

TEST COMPLETE
